+++ internal/campaigns/features.go
@@ -11,6 +11,7 @@ type featureFlags struct {
 	allowArrayEnvironments   bool
 	includeAutoAuthorDetails bool
 	useGzipCompression       bool
+	allowtransformChanges    bool
 }
 
 func (ff *featureFlags) setFromVersion(version string) error {
@@ -22,6 +23,7 @@ func (ff *featureFlags) setFromVersion(version string) error {
 		{&ff.allowArrayEnvironments, ">= 3.23.0", "2020-11-24"},
 		{&ff.includeAutoAuthorDetails, ">= 3.20.0", "2020-09-10"},
 		{&ff.useGzipCompression, ">= 3.21.0", "2020-10-12"},
+		{&ff.allowtransformChanges, ">= 3.23.0", "2020-12-11"},
 	} {
 		value, err := api.CheckSourcegraphVersion(version, feature.constraint, feature.minDate)
 		if err != nil {